Rework building registry to extra class
