+++ src/api/java/com/minecolonies/api/util/constant/BuildingConstants.java
@@ -1,0 +1,32 @@
+package com.minecolonies.api.util.constant;
+
+/**
+ * Constants regarding buildings.
+ */
+public final class BuildingConstants
+{
+    /**
+     * The default mac building level.
+     */
+    public static final int CONST_DEFAULT_MAX_BUILDING_LEVEL = 5;
+
+    /**
+     * Tag if the building has no workOrder.
+     */
+    public static final int NO_WORK_ORDER = 0;
+
+    /**
+     * Max priority of a building.
+     */
+    public static final int MAX_PRIO = 10;
+
+    /**
+     * Private constructor to hide implicit public one.
+     */
+    private BuildingConstants()
+    {
+        /**
+         * Intentionally left empty.
+         */
+    }
+}
