+++ src/main/java/com/minecolonies/coremod/colony/buildings/BuildingDeliveryman.java
@@ -18,6 +18,8 @@ import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.math.BlockPos;
 import org.jetbrains.annotations.NotNull;
 
+import static com.minecolonies.api.util.constant.BuildingConstants.CONST_DEFAULT_MAX_BUILDING_LEVEL;
+
 /**
  * Class of the warehouse building.
  */
